修改release下编译问题

+++ muduo/base/TimeZone.cc
@@ -168,7 +168,8 @@ class File : noncopyable
 // RFC 8536: https://www.rfc-editor.org/rfc/rfc8536.html
 bool readDataBlock(File& f, struct TimeZone::Data* data, bool v1)
 {
-  const int time_size = v1 ? sizeof(int32_t) : sizeof(int64_t);
+    // error: conversion to ‘int’ from ‘long unsigned int’ may alter its value [-Werror=conversion]
+  const int time_size = v1 ? static_cast<int>(sizeof(int32_t)) : static_cast<int>(sizeof(int64_t));
   const int32_t isutccnt = f.readInt32();
   const int32_t isstdcnt = f.readInt32();
   const int32_t leapcnt = f.readInt32();
